+++ modules/Custom/Menu.php
@@ -1,4 +1,11 @@
 <?php
+/**
+ * Custom module Menu entries
+ *
+ * @uses $menu global var
+ *
+ * @see  Menu.php in root folder
+ */
 
 // Custom Students programs
 if ( $RosarioModules['Students'] )
@@ -17,15 +24,17 @@ if ( $RosarioModules['Students'] )
 
 // Custom Users programs
 if ( $RosarioModules['Users'] )
+{
 	$menu['Users']['admin'] += array(
 		3 => _( 'Utilities' ),
 		'Custom/NotifyParents.php' => _( 'Notify Parents' ),
 	);
+}
 
 // Custom Attendance programs
 if ( $RosarioModules['Attendance'] )
+{
 	$menu['Attendance']['admin'] += array(
 		'Custom/AttendanceSummary.php' => _( 'Attendance Summary' )
 	);
-
-?>
+}