+++ modules/Grades/Menu.php
@@ -1,5 +1,14 @@
 <?php
+/**
+ * Grades module Menu entries
+ *
+ * @uses $menu global var
+ *
+ * @see  Menu.php in root folder
+ */
+
 $menu['Grades']['admin'] = array(
+	'title' => _( 'Grades' ),
 	'default' => 'Grades/GPARankList.php',
 	'Grades/ReportCards.php' => _( 'Report Cards' ),
 	'Grades/HonorRoll.php' => _( 'Honor Roll' ),
@@ -28,6 +37,7 @@ $menu['Grades']['admin'] = array(
 );
 
 $menu['Grades']['teacher'] = array(
+	'title' => _( 'Grades' ),
 	'default' => 'Grades/Grades.php',
 	'Grades/InputFinalGrades.php' => _( 'Input Final Grades' ),
 	'Grades/ReportCards.php' => _( 'Report Cards' ),
@@ -51,6 +61,7 @@ $menu['Grades']['teacher'] = array(
 );
 
 $menu['Grades']['parent'] = array(
+	'title' => _( 'Grades' ),
 	'default' => 'Grades/StudentGrades.php',
 	'Grades/StudentGrades.php' => _( 'Gradebook Grades' ),
 	'Grades/FinalGrades.php' => _( 'Final Grades' ),
@@ -60,14 +71,14 @@ $menu['Grades']['parent'] = array(
 );
 
 if ( $RosarioModules['Users'] )
+{
 	$menu['Users']['admin'] += array(
 		'Users/TeacherPrograms.php&include=Grades/InputFinalGrades.php' => _( 'Input Final Grades' ),
 		'Users/TeacherPrograms.php&include=Grades/Grades.php' => _( 'Gradebook Grades' ),
 		'Users/TeacherPrograms.php&include=Grades/AnomalousGrades.php' => _( 'Anomalous Grades' )
 	);
+}
 
 $exceptions['Grades'] = array(
 	'Grades/CalcGPA.php' => true
 );
-
-?>