+++ CHANGES.md
@@ -29,6 +29,7 @@ Changes in 2.9
 - Remove functions/IncludeOnce.php file
 - Add all jqPlot plugins & excanvas (IE version<9 compat) in assets/js/jqplot
 - Regroup Charts (jqPlot) functions in ProgramFunctions/Charts.fnc.php
+- Add Module Title to Menu.php files
 
 Changes in 2.8.11
 -----------------